Rope regeneration, refactored code
Created custom editor for RopeController, which allows the rope to be regenerated with new parameters.
Refactored RopeController class to allow regeneration. Regenerate & Create functions now independently set up the beginning of the rope, then call the AddSegments function to procedurally generate the rope. RopeController now also stores a reference to the first segment of the rope. As this is the piece which the RopeController script is assigned to when created from the editor, this is not deleted when regenerating the rope.
RopeController no longer uses lineRenderer variables introduced in 5.6.1f1, in order to keep compatability. 
RopeController now attempts to auto-assign the rope material via a Resources lookup if no material has been assigned.